Fixed an error in get_statistic

+++ tools/graph_reduce.cpp
@@ -51,7 +51,7 @@ void GraphReduce(const std::string input_pt, const std::string output_pt,
     });
   }
   finish_cv.wait(lck, [&] { return pending_packages.load() == 0; });
-  max_vid_atom.store((int(max_vid_atom.load() / 64) + 1) * 64);
+  max_vid_atom.store((size_t(max_vid_atom.load() / 64) + 1) * 64);
   LOG_INFO("#maximum vid: ", max_vid_atom.load());
   // delete doc;
   VID_T* vid_map = (VID_T*)malloc(sizeof(VID_T) * max_vid_atom.load());
